Tasked done:  [+] Fixed routes  [+] Moved secret_key to .env file
