+++ backend/.bin/dev.js
@@ -1,0 +1,2 @@
+require('dotenv/config')
+require('./../index')
