+++ backend/index.js
@@ -5,9 +5,11 @@ const bodyParser = require('body-parser')
 const controller = require('./module/controller')
 const middleWare = require('./middleware')
 
-const PORT = 3001
+const PORT = process.env.PORT || 3000
+
 const server = jsonServer.create()
-const router = jsonServer.router('./backend/data.json')
+const router = jsonServer.router('./backend/json/db.json');
+
 
 
 //this is neaded to use post, patch etc etc
@@ -17,7 +19,9 @@ server.use(jsonServer.defaults()); //load default jsonServer
 server.post('/login', controller.login)
 
 server.use(middleWare) 
+
 server.use('/api',router)
+
 
 server.listen(PORT, ()=>{
     console.log(`Now listening ( ͡°╭͜ʖ╮͡° )  @ ${PORT}`)
