+++ backend/json/test_db2.json
@@ -1,0 +1,119 @@
+{
+  "users": [
+    {
+      "id": 1,
+      "username": "dev-amor",
+      "password": "12345"
+    },
+    {
+      "id": 2,
+      "username": "dev-aquino",
+      "password": "12345"
+    },
+    {
+      "id": 3,
+      "username": "dev-astillero",
+      "password": "12345"
+    },
+    {
+      "id": 4,
+      "username": "dev-atento",
+      "password": "12345"
+    },
+    {
+      "id": 5,
+      "username": "dev-azores",
+      "password": "12345"
+    },
+    {
+      "id": 6,
+      "username": "dev-balbedina",
+      "password": "12345"
+    },
+    {
+      "id": 7,
+      "username": "dev-bandagosa",
+      "password": "12345"
+    },
+    {
+      "id": 8,
+      "username": "dev-camba",
+      "password": "12345"
+    },
+    {
+      "id": 9,
+      "username": "dev-danas",
+      "password": "12345"
+    },
+    {
+      "id": 10,
+      "username": "dev-delosreyes",
+      "password": "12345"
+    },
+    {
+      "id": 11,
+      "username": "dev-dulva",
+      "password": "12345"
+    },
+    {
+      "id": 12,
+      "username": "dev-frecia",
+      "password": "12345"
+    },
+    {
+      "id": 13,
+      "username": "dev-ifurung",
+      "password": "12345"
+    },
+    {
+      "id": 14,
+      "username": "dev-ludovice",
+      "password": "12345"
+    },
+    {
+      "id": 15,
+      "username": "dev-marbella",
+      "password": "12345"
+    },
+    {
+      "id": 16,
+      "username": "dev-medes",
+      "password": "12345"
+    },
+    {
+      "id": 17,
+      "username": "dev-menor",
+      "password": "12345"
+    },
+    {
+      "id": 18,
+      "username": "dev-mirabete",
+      "password": "12345"
+    },
+    {
+      "id": 19,
+      "username": "dev-molleno",
+      "password": "12345"
+    },
+    {
+      "id": 20,
+      "username": "dev-norte",
+      "password": "12345"
+    },
+    {
+      "id": 21,
+      "username": "dev-restum",
+      "password": "12345"
+    },
+    {
+      "id": 22,
+      "username": "dev-raquion",
+      "password": "12345"
+    },
+    {
+      "id": 23,
+      "username": "dev-serra",
+      "password": "12345"
+    }
+  ]
+}
